feat(agent): anticipatory per-view greeting — judge intent + live-state, transcripts scope (#13587)
Extract live-state renderers into leaf module page-scoped-live-state.ts (shared
by page-scoped-context provider and the proactive decider, no cycle). Add
page-transcripts scope + renderer. Rework judge to greet from declared intent
grounded in live state; drop confidence floor only for intent-bearing views.
Emit anticipatoryIntent + viewPurpose on VIEW_SWITCHED.
