+++ packages/agent/src/providers/page-scoped-context.ts
@@ -18,14 +18,6 @@ import type {
   UUID,
 } from "@elizaos/core";
 import { stringToUuid } from "@elizaos/core";
-import type {
-  AppRunSummary,
-  RegistryAppInfo,
-  WalletBalancesResponse,
-  WalletConfigStatus,
-  WalletNftsResponse,
-  WalletTradingProfileResponse,
-} from "@elizaos/shared";
 import {
   extractConversationMetadataFromRoom,
   isPageScopedConversationMetadata,
@@ -35,6 +27,7 @@ import {
   formatRelativeTimestamp,
   formatSpeakerLabel,
 } from "../shared/conversation-format.ts";
+import { renderLiveStateForScope } from "./page-scoped-live-state.ts";
 
 const SOURCE_TAIL_LIMIT = 6;
 const SOURCE_TAIL_MIN_FOR_INCLUSION = 2;
@@ -63,6 +56,8 @@ const PAGE_SCOPE_BRIEF: Record<string, string> = {
     "The user is in the Wallet view. They can inspect token inventory, NFTs, LP position status, current balance, P&L, activity, EVM/Solana addresses, RPC/provider readiness, wallet/RPC settings, and native Hyperliquid and Polymarket readiness. There are no chain filters in this surface. When the user asks what to do, recommend the smallest concrete wallet action and confirm asset/market, amount, destination/outcome, slippage/risk limits, and execution path before invoking any available action. If the user asks about Hyperliquid or Polymarket, prefer native app surfaces for reads/status. Never invent balances, positions, fills, markets, odds, or execution support.",
   "page-knowledge":
     "The user is in the Knowledge view. This surface lists knowledge records, including chat attachments ingested by room, sender, sender role, and media format (image, audio, video, pdf, text, transcript, file). Owner/DM-chat knowledge is owner-private and never surfaces into public rooms; public/community-room attachments are scoped to their sender. The user can browse, search, retag, rescope (owner-only), and delete knowledge items; deleting the sole referent of a media file makes those bytes garbage-collectible. When the user asks what to do, ground answers in the live knowledge state below (recent-attachment counts by source and format) and recommend the smallest concrete filing, retag, rescope, or cleanup action. Never invent knowledge records, media, tags, or scopes.",
+  "page-transcripts":
+    "The user is in the Transcripts view. This surface lists recorded voice transcripts they can play, scrub, and read with word-level sync. The user can review a transcript, ask for a summary or action items, or search across recordings. When the user asks what to do, ground answers in the live transcript state below (recent-transcript count and recency) and offer to summarize or extract action items from the most recent recording. Never invent transcripts, speakers, or contents.",
   "automation-draft":
     "This is an automation-creation room. The user wants to create exactly one automation. Decide the right shape based on their description and call the matching action exactly once:\n" +
     '- Recurring prompt or scheduled instruction (e.g. "every morning summarize my inbox") → TRIGGER with op="create" specifying displayName, instructions, and schedule (interval/once/cron). The agent transparently materializes a single-node workflow that runs the instructions when the trigger fires.\n' +
@@ -148,542 +143,6 @@ async function fetchSourceTail(
     ageLabel: formatRelativeTimestamp(mem.createdAt),
     role: inferRole(mem, runtime.agentId),
   }));
-}
-
-async function renderCharacterLiveState(
-  runtime: IAgentRuntime,
-): Promise<string | null> {
-  const character = runtime.character;
-  if (!character) return null;
-  const lines: string[] = ["Live character state:"];
-  lines.push(`- Name: ${character.name ?? "(unnamed)"}`);
-  const bio = (character as { bio?: unknown }).bio;
-  if (typeof bio === "string" && bio.trim().length > 0) {
-    lines.push(`- Bio: ${bio.trim().slice(0, 200)}`);
-  } else if (Array.isArray(bio) && bio.length > 0) {
-    lines.push(`- Bio entries: ${bio.length}`);
-  }
-  const exampleCount = Array.isArray(character.messageExamples)
-    ? character.messageExamples.length
-    : 0;
-  lines.push(`- Message examples: ${exampleCount}`);
-  return lines.join("\n");
-}
-
-interface BrowserBridgeCompanionLiveStatus {
-  connectionState: string;
-  browser: string;
-  profileLabel?: string | null;
-  extensionVersion?: string | null;
-}
-
-function getLocalApiUrls(path: string): string[] {
-  const normalizedPath = path.startsWith("/") ? path : `/${path}`;
-  const configuredPort = process.env.API_PORT || process.env.SERVER_PORT;
-  const ports = configuredPort
-    ? [configuredPort, configuredPort === "31337" ? "2138" : "31337"]
-    : ["2138", "31337"];
-  return [...new Set(ports)].map(
-    (port) => `http://127.0.0.1:${port}${normalizedPath}`,
-  );
-}
-
-async function fetchLocalJson<T>(
-  path: string,
-  timeoutMs = 1500,
-): Promise<T | null> {
-  for (const url of getLocalApiUrls(path)) {
-    try {
-      const response = await fetch(url, {
-        signal: AbortSignal.timeout(timeoutMs),
-      });
-      if (!response.ok) continue;
-      return (await response.json()) as T;
-    } catch {
-      // error-policy:J4 local-API port failover — the dev API lives on one of a
-      // small set of candidate ports; a connection failure means "try the next
-      // port". Total exhaustion returns null, which every caller renders as an
-      // explicit "unavailable from the … API" line (agent-visible, not silence).
-      continue;
-    }
-  }
-  return null;
-}
-
-async function fetchBrowserBridgeCompanionLiveStatus(): Promise<
-  BrowserBridgeCompanionLiveStatus[] | null
-> {
-  const payload = await fetchLocalJson<{
-    companions?: Array<{
-      connectionState?: string;
-      browser?: string;
-      profileLabel?: string | null;
-      extensionVersion?: string | null;
-    }>;
-  }>("/api/browser-bridge/companions");
-  if (!payload) return null;
-  if (!Array.isArray(payload.companions)) return [];
-  return payload.companions.map((companion) => ({
-    connectionState: companion.connectionState ?? "unknown",
-    browser: companion.browser ?? "chrome",
-    profileLabel: companion.profileLabel ?? null,
-    extensionVersion: companion.extensionVersion ?? null,
-  }));
-}
-
-/**
- * Minimal structural view of the browser workspace service. Typed locally so
- * the provider never takes an import edge into the browser plugin; the
- * service is resolved by its runtime service type and the section is omitted
- * for agents that do not have the plugin enabled.
- */
-interface BrowserWorkspaceSnapshotView {
-  mode: string;
-  tabs: Array<{ title?: string | null; url: string; visible?: boolean }>;
-}
-interface BrowserWorkspaceServiceLike {
-  getWorkspaceSnapshot(): Promise<BrowserWorkspaceSnapshotView>;
-}
-
-const BROWSER_SERVICE_TYPE = "browser";
-
-function isBrowserWorkspaceService(
-  service: unknown,
-): service is BrowserWorkspaceServiceLike {
-  return (
-    typeof (service as { getWorkspaceSnapshot?: unknown } | null)
-      ?.getWorkspaceSnapshot === "function"
-  );
-}
-
-async function renderBrowserLiveState(
-  runtime: IAgentRuntime,
-): Promise<string | null> {
-  const service = runtime.getService(BROWSER_SERVICE_TYPE);
-  if (!isBrowserWorkspaceService(service)) return null;
-  try {
-    const snapshot = await service.getWorkspaceSnapshot();
-    const lines: string[] = [
-      `Live browser state: bridge=${snapshot.mode}, ${snapshot.tabs.length} tab${snapshot.tabs.length === 1 ? "" : "s"}.`,
-    ];
-    for (const tab of snapshot.tabs.slice(0, 6)) {
-      const flags = tab.visible ? "[visible]" : "";
-      lines.push(`- ${tab.title || "(untitled)"} — ${tab.url} ${flags}`.trim());
-    }
-
-    // Agent Browser Bridge companion status — so the agent can tell the user
-    // to install the extension when it isn't connected and reference the
-    // connected profile accurately when it is.
-    const companions = await fetchBrowserBridgeCompanionLiveStatus();
-    if (companions === null) {
-      lines.push(
-        "Agent Browser Bridge companion: status unknown (companion API unreachable).",
-      );
-    } else if (companions.length === 0) {
-      lines.push(
-        "Agent Browser Bridge companion: not installed — tell the user to click 'Install Agent Browser Bridge' in the chat panel to build the extension and load it into Chrome.",
-      );
-    } else {
-      const connected = companions.filter(
-        (companion) => companion.connectionState === "connected",
-      );
-      if (connected.length === 0) {
-        lines.push(
-          "Agent Browser Bridge companion: extension present but not connected — ask the user to open the Agent Browser Bridge extension in Chrome so it can pair.",
-        );
-      } else {
-        lines.push(
-          `Agent Browser Bridge companion: connected (${connected.length} profile${connected.length === 1 ? "" : "s"}).`,
-        );
-        for (const companion of connected.slice(0, 3)) {
-          const browser = companion.browser === "safari" ? "Safari" : "Chrome";
-          const profile = companion.profileLabel?.trim() || "Default";
-          const version = companion.extensionVersion
-            ? ` v${companion.extensionVersion}`
-            : "";
-          lines.push(`- ${browser} / ${profile}${version}`);
-        }
-      }
-    }
-
-    return lines.join("\n");
-  } catch (err) {
-    // error-policy:J4 explicit user-facing degrade — one failing subsection
-    // (browser service throwing) must not blank the whole provider, so it
-    // degrades to an omitted section; the failure is reported so it reaches the
-    // RECENT_ERRORS provider instead of vanishing.
-    runtime.reportError("PageScopedContext.browserLiveState", err);
-    return null;
-  }
-}
-
-function dedupeApps(
-  groups: Array<RegistryAppInfo[] | null>,
-): RegistryAppInfo[] {
-  const apps = new Map<string, RegistryAppInfo>();
-  for (const group of groups) {
-    if (!group) continue;
-    for (const app of group) {
-      if (!app.name || apps.has(app.name)) continue;
-      apps.set(app.name, app);
-    }
-  }
-  return [...apps.values()].sort((left, right) =>
-    left.displayName.localeCompare(right.displayName),
-  );
-}
-
-async function renderAppsLiveState(): Promise<string | null> {
-  const [catalogApps, serverApps, runs] = await Promise.all([
-    fetchLocalJson<RegistryAppInfo[]>("/api/catalog/apps"),
-    fetchLocalJson<RegistryAppInfo[]>("/api/apps"),
-    fetchLocalJson<AppRunSummary[]>("/api/apps/runs"),
-  ]);
-  if (!catalogApps && !serverApps && !runs) {
-    return "Live apps state: unavailable from the Apps API.";
-  }
-
-  const apps = dedupeApps([catalogApps, serverApps]);
-  const activeRuns = runs ?? [];
-  const lines: string[] = [
-    `Live apps state: ${apps.length} catalog app${apps.length === 1 ? "" : "s"}, ${activeRuns.length} running app${activeRuns.length === 1 ? "" : "s"}.`,
-  ];
-
-  if (activeRuns.length > 0) {
-    lines.push("Running apps:");
-    for (const run of activeRuns.slice(0, 8)) {
-      const health = run.health.state ? ` health=${run.health.state}` : "";
-      const viewer = run.viewerAttachment
-        ? ` viewer=${run.viewerAttachment}`
-        : "";
-      const summary = run.summary ? ` — ${run.summary.slice(0, 140)}` : "";
-      lines.push(
-        `- ${run.displayName} (${run.appName}) status=${run.status}${health}${viewer}${summary}`,
-      );
-    }
-  } else {
-    lines.push("Running apps: none.");
-  }
-
-  if (apps.length > 0) {
-    lines.push("Catalog sample:");
-    for (const app of apps.slice(0, 12)) {
-      const capabilities =
-        app.capabilities.length > 0
-          ? ` capabilities=${app.capabilities.slice(0, 4).join(", ")}`
-          : "";
-      lines.push(
-        `- ${app.displayName} (${app.name}) category=${app.category}${capabilities}`,
-      );
-    }
-  }
-
-  return lines.join("\n");
-}
-
-function shortAddress(address: string | null | undefined): string {
-  if (!address) return "(not configured)";
-  if (address.length <= 14) return address;
-  return `${address.slice(0, 6)}...${address.slice(-4)}`;
-}
-
-function readyLabel(value: boolean | undefined): string {
-  if (value === true) return "ready";
-  if (value === false) return "not ready";
-  return "unknown";
-}
-
-function hasPositiveAmount(value: string | null | undefined): boolean {
-  if (!value) return false;
-  const parsed = Number.parseFloat(value);
-  return Number.isFinite(parsed) && parsed > 0;
-}
-
-interface HyperliquidLiveStatus {
-  publicReadReady?: boolean;
-  signerReady?: boolean;
-  executionReady?: boolean;
-  executionBlockedReason?: string | null;
-  accountAddress?: string | null;
-}
-
-interface PolymarketLiveStatus {
-  publicReads?: {
-    ready?: boolean;
-  };
-  trading?: {
-    ready?: boolean;
-    credentialsReady?: boolean;
-    reason?: string | null;
-    missing?: readonly string[];
-  };
-}
-
-async function renderWalletLiveState(): Promise<string | null> {
-  const [config, balances, nfts, profile, hyperliquidStatus, polymarketStatus] =
-    await Promise.all([
-      fetchLocalJson<WalletConfigStatus>("/api/wallet/config"),
-      fetchLocalJson<WalletBalancesResponse>("/api/wallet/balances"),
-      fetchLocalJson<WalletNftsResponse>("/api/wallet/nfts"),
-      fetchLocalJson<WalletTradingProfileResponse>(
-        "/api/wallet/trading/profile?window=24h&source=all",
-      ),
-      fetchLocalJson<HyperliquidLiveStatus>("/api/hyperliquid/status"),
-      fetchLocalJson<PolymarketLiveStatus>("/api/polymarket/status"),
-    ]);
-
-  if (
-    !config &&
-    !balances &&
-    !nfts &&
-    !profile &&
-    !hyperliquidStatus &&
-    !polymarketStatus
-  ) {
-    return "Live wallet state: unavailable from the Wallet API.";
-  }
-
-  const lines: string[] = ["Live wallet state:"];
-  if (config) {
-    lines.push(`- Wallet source: ${config.walletSource ?? "unknown"}`);
-    lines.push(`- EVM address: ${shortAddress(config.evmAddress)}`);
-    lines.push(`- Solana address: ${shortAddress(config.solanaAddress)}`);
-    lines.push(
-      `- RPC providers: EVM=${config.selectedRpcProviders.evm}, Solana=${config.selectedRpcProviders.solana}`,
-    );
-    lines.push(
-      `- Readiness: EVM balances ${readyLabel(config.evmBalanceReady)}, Solana balances ${readyLabel(config.solanaBalanceReady)}, execution ${readyLabel(config.executionReady)}`,
-    );
-    if (config.executionBlockedReason) {
-      lines.push(`- Execution blocked: ${config.executionBlockedReason}`);
-    }
-    lines.push(
-      `- Signing: EVM=${config.evmSigningCapability ?? "unknown"}, Solana=${readyLabel(config.solanaSigningAvailable)}`,
-    );
-  }
-
-  const assetLines: string[] = [];
-  if (balances?.evm) {
-    for (const chain of balances.evm.chains) {
-      if (hasPositiveAmount(chain.nativeBalance)) {
-        assetLines.push(`${chain.nativeBalance} ${chain.nativeSymbol}`);
-      }
-      for (const token of chain.tokens) {
-        if (hasPositiveAmount(token.balance)) {
-          assetLines.push(`${token.balance} ${token.symbol}`);
-        }
-      }
-    }
-  }
-  if (balances?.solana) {
-    if (hasPositiveAmount(balances.solana.solBalance)) {
-      assetLines.push(`${balances.solana.solBalance} SOL`);
-    }
-    for (const token of balances.solana.tokens) {
-      if (hasPositiveAmount(token.balance)) {
-        assetLines.push(`${token.balance} ${token.symbol}`);
-      }
-    }
-  }
-  if (balances?.evm || balances?.solana) {
-    lines.push(
-      `- Token inventory: ${assetLines.length} asset${assetLines.length === 1 ? "" : "s"}.`,
-    );
-    for (const asset of assetLines.slice(0, 10)) {
-      lines.push(`  - ${asset}`);
-    }
-  }
-
-  if (nfts) {
-    const evmNftCount = nfts.evm.reduce(
-      (sum, chain) => sum + chain.nfts.length,
-      0,
-    );
-    const solanaNftCount = nfts.solana?.nfts.length ?? 0;
-    const nftCount = evmNftCount + solanaNftCount;
-    lines.push(`- NFTs: ${nftCount} item${nftCount === 1 ? "" : "s"}.`);
-  }
-
-  if (profile) {
-    lines.push(
-      `- 24h activity: ${profile.summary.totalSwaps} swap${profile.summary.totalSwaps === 1 ? "" : "s"}, realized P&L ${profile.summary.realizedPnlBnb} BNB, volume ${profile.summary.volumeBnb} BNB.`,
-    );
-  }
-
-  if (hyperliquidStatus) {
-    lines.push(
-      `- Hyperliquid native: public reads ${readyLabel(hyperliquidStatus.publicReadReady)}, signer ${readyLabel(hyperliquidStatus.signerReady)}, execution ${readyLabel(hyperliquidStatus.executionReady)}, account ${shortAddress(hyperliquidStatus.accountAddress)}.`,
-    );
-    if (hyperliquidStatus.executionBlockedReason) {
-      lines.push(
-        `- Hyperliquid execution blocked: ${hyperliquidStatus.executionBlockedReason}`,
-      );
-    }
-  }
-
-  if (polymarketStatus) {
-    lines.push(
-      `- Polymarket native: public reads ${readyLabel(polymarketStatus.publicReads?.ready)}, credentials ${readyLabel(polymarketStatus.trading?.credentialsReady)}, trading ${readyLabel(polymarketStatus.trading?.ready)}.`,
-    );
-    if (polymarketStatus.trading?.reason) {
-      lines.push(
-        `- Polymarket trading blocked: ${polymarketStatus.trading.reason}`,
-      );
-    }
-  }
-
-  return lines.join("\n");
-}
-
-async function renderAutomationsLiveState(
-  runtime: IAgentRuntime,
-): Promise<string | null> {
-  try {
-    const tasks = await runtime.getTasks({ agentIds: [runtime.agentId] });
-    if (tasks.length === 0) return "Live automations state: no tasks defined.";
-    const lines: string[] = [
-      `Live automations state: ${tasks.length} task${tasks.length === 1 ? "" : "s"}.`,
-    ];
-    for (const task of tasks.slice(0, 8)) {
-      const name = task.name;
-      const tagList =
-        Array.isArray(task.tags) && task.tags.length > 0
-          ? ` [${task.tags.join(", ")}]`
-          : "";
-      lines.push(`- ${name}${tagList}`);
-    }
-    return lines.join("\n");
-  } catch (err) {
-    // error-policy:J4 explicit user-facing degrade — a failing task read must
-    // not blank the whole provider; the automations section is omitted and the
-    // failure is reported so it surfaces via the RECENT_ERRORS provider.
-    runtime.reportError("PageScopedContext.automationsLiveState", err);
-    return null;
-  }
-}
-
-const KNOWLEDGE_LIVE_STATE_WINDOW_MS = 7 * 24 * 60 * 60 * 1000;
-const KNOWLEDGE_LIVE_STATE_SCAN_LIMIT = 500;
-const ATTACHMENT_DOCUMENT_TAG = "attachment";
-const MEDIA_FORMAT_TAG_PREFIX = "media-format:";
-const TRANSCRIPT_DOCUMENT_TAG = "transcript";
-
-function documentStringTags(
-  metadata: Record<string, unknown> | undefined,
-): string[] {
-  const tags = metadata?.tags;
-  return Array.isArray(tags)
-    ? tags.filter((value): value is string => typeof value === "string")
-    : [];
-}
-
-function documentMediaFormat(
-  metadata: Record<string, unknown> | undefined,
-  tags: string[],
-): string {
-  if (typeof metadata?.mediaFormat === "string" && metadata.mediaFormat) {
-    return metadata.mediaFormat;
-  }
-  const tagged = tags.find((tag) => tag.startsWith(MEDIA_FORMAT_TAG_PREFIX));
-  return tagged ? tagged.slice(MEDIA_FORMAT_TAG_PREFIX.length) : "file";
-}
-
-/**
- * Live knowledge state for the Knowledge view (#13593): counts of recently
- * ingested chat attachments (and transcript mirrors) over the trailing week,
- * broken down by media format. This is the count surface slice 2's proactive
- * greeting reads ("You have N new attachments from Discord this week…").
- */
-async function renderKnowledgeLiveState(
-  runtime: IAgentRuntime,
-): Promise<string | null> {
-  const now = Date.now();
-  const windowStart = now - KNOWLEDGE_LIVE_STATE_WINDOW_MS;
-  let recentAttachments = 0;
-  let recentTranscripts = 0;
-  const byFormat = new Map<string, number>();
-
-  try {
-    const batch = await runtime.getMemories({
-      tableName: "documents",
-      // Scope to THIS agent so a shared/multi-agent adapter does not count
-      // another agent's attachment/transcript records into this page's context.
-      agentId: runtime.agentId,
-      count: KNOWLEDGE_LIVE_STATE_SCAN_LIMIT,
-      offset: 0,
-    });
-    for (const memory of batch) {
-      if (memory.agentId && memory.agentId !== runtime.agentId) continue;
-      const metadata = (memory.metadata ?? undefined) as
-        | Record<string, unknown>
-        | undefined;
-      const tags = documentStringTags(metadata);
-      const isAttachment = tags.includes(ATTACHMENT_DOCUMENT_TAG);
-      const isTranscript = tags.includes(TRANSCRIPT_DOCUMENT_TAG);
-      if (!isAttachment && !isTranscript) continue;
-      const at =
-        (typeof metadata?.addedAt === "number"
-          ? metadata.addedAt
-          : undefined) ??
-        (typeof metadata?.timestamp === "number"
-          ? metadata.timestamp
-          : undefined) ??
-        (typeof memory.createdAt === "number" ? memory.createdAt : undefined);
-      if (typeof at === "number" && at < windowStart) continue;
-      if (isAttachment) {
-        recentAttachments += 1;
-        const format = documentMediaFormat(metadata, tags);
-        byFormat.set(format, (byFormat.get(format) ?? 0) + 1);
-      } else if (isTranscript) {
-        recentTranscripts += 1;
-      }
-    }
-  } catch (err) {
-    runtime.reportError("PageScopedContext.knowledgeLiveState", err);
-    return "Live knowledge state: unavailable (documents store unreachable).";
-  }
-
-  const lines: string[] = [
-    `Live knowledge state (last 7 days): ${recentAttachments} ingested chat attachment${
-      recentAttachments === 1 ? "" : "s"
-    }, ${recentTranscripts} transcript mirror${
-      recentTranscripts === 1 ? "" : "s"
-    }.`,
-  ];
-  if (byFormat.size > 0) {
-    const parts = [...byFormat.entries()]
-      .sort((a, b) => b[1] - a[1])
-      .map(([format, count]) => `${format}=${count}`);
-    lines.push(`Attachments by format: ${parts.join(", ")}.`);
-  }
-  return lines.join("\n");
-}
-
-async function renderLiveStateForScope(
-  runtime: IAgentRuntime,
-  scope: ConversationScope,
-): Promise<string | null> {
-  switch (scope) {
-    case "page-character":
-      return renderCharacterLiveState(runtime);
-    case "page-knowledge":
-      return renderKnowledgeLiveState(runtime);
-    case "page-browser":
-      return renderBrowserLiveState(runtime);
-    case "page-automations":
-    case "automation-draft":
-      return renderAutomationsLiveState(runtime);
-    case "page-apps":
-      return renderAppsLiveState();
-    case "page-connectors":
-    case "page-plugins":
-    case "page-settings":
-      return null;
-    case "page-wallet":
-      return renderWalletLiveState();
-    default:
-      return null;
-  }
 }
 
 function formatSourceTail(entries: SourceTailEntry[]): string {
