+++ packages/shared/src/contracts/__tests__/schema-vs-type-drift.test.ts
@@ -42,6 +42,7 @@ const VALID_CONVERSATION_SCOPES = new Set([
   "page-browser",
   "page-automations",
   "page-knowledge",
+  "page-transcripts",
 ]);
 
 const VALID_CONVERSATION_AUTOMATION_TYPES = new Set([
